+++ .github/workflows/main.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Check formatting, linting and import sorting
         run: ./scripts/setup_functions check
 
-#  test:
+#  test: TODO: Comment back in when tests done
 #    needs: [check]
 #    runs-on: ubuntu-latest
 #
@@ -28,7 +28,7 @@ jobs:
 #      - name: Run tests
 ##       env:  # Add environment variables required for tests
 #        run: |
-#          pytest
+#          just test
 
   lint-dockerfile:
     runs-on: ubuntu-latest
