+++ actions/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 # Register your models here.
-from .models import Actions
+from .models import Action
 
 
-admin.site.register(Actions)
+admin.site.register(Action)
