+++ actions/templates/actions/detail.html
@@ -6,7 +6,7 @@
 </head>
 <body>
     {% block content %}
-        <h2>{{ action.action_name}}</h2>
+        <h2>{{ action.name}}</h2>
         {{ action.description}}
     {% endblock %}
 </body>
