+++ registry/settings.py
